Removed screen in favor of a new state module and game loop
